backend/requirements.txt
streamlit==1.40.0
psycopg2-binary==2.9.10